Setup routing using vue-router in frontend
