+++ frontend/package.json
@@ -13,7 +13,8 @@
     "stylus": "^0.59.0",
     "stylus-loader": "^7.1.0",
     "vite-plugin-string-replace": "^1.0.2",
-    "vue": "^3.2.41"
+    "vue": "^3.2.41",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^3.2.0",
