+++ frontend/src/main.js
@@ -1,7 +1,18 @@
 import { createApp } from 'vue'
+import { createRouter, createWebHistory } from 'vue-router'
+import routes from './routes'
 import { GlobalWorkerOptions } from 'pdfjs-dist'
 import './style.css'
 import App from './App.vue'
 
+const router = createRouter({
+  history: createWebHistory(),
+  routes
+})
+
 GlobalWorkerOptions.workerSrc = '/pdf.worker.js'
-createApp(App).mount('#app')
+const app = createApp(App)
+
+app.use(router)
+
+app.mount('#app')
