+++ frontend/src/routes.js
@@ -1,0 +1,11 @@
+import Home from './pages/Home.vue'
+
+const routes = [
+  {
+    path: '/',
+    name: 'home',
+    component: Home
+  }
+]
+
+export default routes
